move `check_and_delete_old_databases` to state
